+++ tests/test.nim
@@ -20,3 +20,5 @@ import ./t_atomics_dsl
 import ./t_thread_id
 import ./t_item_processing
 import ./t_lockfree_stack_typestates
+import ./t_backoff
+import ./t_bind_client